core/standards/XMLHttpRequest/http-auth/http-auth-change-userinfo.htm

<!doctype html>
<meta charset=utf-8>
<title>Change HTTP auth username and password using XHR</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<p>Tests require uncluttered HTTP auth.</p>
<div id=log></div>

<script>
var t = async_test()
var xhr = new XMLHttpRequest()
var time = new Date().getTime()

xhr.open('GET', 'resources/login.php?' + time, true, 'yo_man', 'boomblaster+' + time + '_1')
xhr.onload = t.step_func(function() {
    assert_equals(xhr.response, 'yo_man:boomblaster+' + time + '_1', 'first response')

    xhr.open('GET', 'resources/login.php?' + time, true, 'yo_man_2', 'boomblaster+' + time + '_2')
    xhr.onload = t.step_func(function() {
        assert_equals(xhr.response, 'yo_man_2:boomblaster+' + time + '_2', 'second response')
        t.done()
    })
    xhr.send()

})
xhr.send()
</script>
